Adding a quick 3 way team PER comparison
So I chose GSW, MIA, and PHI (Good, Average, Bad teams respectively) to compare the Average PER for team performance. Player Efficiency Rating (PER) is a stat that shows a players efficiency in offensive and defensive situations, adjusted for league pace (the calculation for PER is serious and can be found w/ a quick google). All data was taken from basketball-reference.com, and exported to excel spreadsheets.

As you can see, the most accurate average appears to be the top 10 PER for top 10 minutes played. This takes into account starters and other players off the bench.
I'm sure we could develop a way to get a more accurate team PER, by weighting a players PER to the amount of minutes play, as opposed to taking the raw top 10. (EX: KD plays 2000 Minutes, and has a PER of 30, his affect is weighted more than someone who plays 1000 Minutes and has a PER of 30).

Feedback on this stat and the idea would be cool.
As far as running calculations on this data, I'm su